+++ packages/cli/src/cli/index.ts
@@ -2,7 +2,7 @@
 
 import pkg from "@pkg";
 import { getRenderCommand } from "@/cli/render";
-import { getCommand } from "@/lib/command";
+import { getCommand } from "@/lib/cli/command";
 import { issues } from "@/lib/issue";
 
 const description = `Dotsecret is designed to simplify the process of fetching secrets from secrets managers and render files with them.
