+++ packages/cli/src/lib/cli/argument.ts
@@ -1,0 +1,9 @@
+import { Argument } from "commander";
+import { parserGlobPatterns } from "./parser";
+
+export type ArgumentTemplates = ReturnType<typeof parserGlobPatterns>;
+
+export const argumentTemplates = new Argument(
+  "[templates]",
+  "Templates glob patterns",
+).argParser(parserGlobPatterns);
