+++ packages/cli/src/lib/cli/command.ts
@@ -1,88 +1,17 @@
 import pkg from "@pkg";
-import { Argument, Command, InvalidArgumentError, Option } from "commander";
-import { LOG_LEVEL_DEFAULT, LOG_LEVELS } from "@/lib/logger";
-import { getAbsolutePath, isFileExists, isPathExists } from "@/lib/utils";
-
-export type ParsedPathExists = string;
-
-export const argParserPathExists = (val: string) => {
-  const absolutePath = getAbsolutePath(val);
-  const pathExists = isPathExists(absolutePath);
-
-  if (!pathExists) {
-    throw new InvalidArgumentError(
-      `The path "${absolutePath}" does not exist.`,
-    );
-  }
-
-  return absolutePath;
-};
-
-export type ParsedFileExists = string;
-
-export const argParserFileExists = (val: string) => {
-  const absolutePath = getAbsolutePath(val);
-  const fileExists = isFileExists(absolutePath);
-
-  if (!fileExists) {
-    throw new InvalidArgumentError(
-      `The file "${absolutePath}" does not exist.`,
-    );
-  }
-
-  return absolutePath;
-};
-
-export type ParsedGlobPatterns = string[];
-
-export const argParserGlobPatterns = (val: string) =>
-  val
-    .split(" ")
-    .map((pattern) => pattern.trim())
-    .filter(Boolean);
-
-export type ParsedOptionConfig = ParsedFileExists;
-
-export const optionConfig = new Option(
-  "-c, --config <path>",
-  "Specify a configuration file",
-).argParser(argParserFileExists);
-
-export type ParsedOptionLogLevel = (typeof LOG_LEVELS)[number];
-
-export const optionLogLevel = new Option("-l, --log-level <level>", "Log level")
-  .choices(LOG_LEVELS)
-  .default(LOG_LEVEL_DEFAULT);
-
-export type ParsedOptionCwd = ParsedPathExists;
-
-export const optionCwd = new Option(
-  "-d, --cwd <path>",
-  "Current working directory",
-)
-  .default(process.cwd())
-  .argParser(argParserPathExists);
-
-export type ParsedOptionForce = boolean;
-
-export const optionForce = new Option(
-  "-f, --force",
-  "Override safety checks and force operations to complete",
-).default(false);
-
-export type ParsedOptionDryRun = boolean;
-
-export const optionDryRun = new Option(
-  "--dry-run",
-  "Show what would be done without actually doing it",
-).default(false);
-
-export type ParsedArgumentTemplates = ParsedGlobPatterns;
-
-export const argumentTemplates = new Argument(
-  "[templates]",
-  "Templates glob patterns",
-).argParser(argParserGlobPatterns);
+import { Command } from "commander";
+import {
+  type OptionConfig,
+  optionConfig,
+  type OptionCwd,
+  optionCwd,
+  type OptionDryRun,
+  optionDryRun,
+  type OptionForce,
+  optionForce,
+  type OptionLogLevel,
+  optionLogLevel,
+} from "./option";
 
 type Options = {
   /** Whether to display the --help option. */
@@ -119,15 +48,15 @@ const DEFAULT_COMMANDS: Commands = {
 
 export type CommandOptions<T = {}> = {
   /** The path to the configuration file. */
-  config?: ParsedOptionConfig;
+  config?: OptionConfig;
   /** Whether to show what would be done without actually doing it. */
-  dryRun: ParsedOptionDryRun;
+  dryRun: OptionDryRun;
   /** The log level. */
-  logLevel: ParsedOptionLogLevel;
+  logLevel: OptionLogLevel;
   /** Whether to override safety checks and force operations to complete. */
-  force: ParsedOptionForce;
+  force: OptionForce;
   /** The current working directory. */
-  cwd: ParsedOptionCwd;
+  cwd: OptionCwd;
 } & T;
 
 export const getCommand = async ({
